database: url -> file_url app: add func add;combine dl and update; add clear settings: change path info, and TAGS to path fix bugs of process_update_info

+++ yandeTime.py
@@ -14,11 +14,7 @@ import re
 from database import ss, Image
 from utils import check_exists
 import os
-from settings import IMG_PATH
-
-TAGS = ['bondage', 'loli', 'pee', 'vibrator',
-        'anal', 'dildo', 'anal_beads', 'masturbation', 'yuri',
-        'cunnilingus', 'fingering', 'pussy_juice', 'fellatio', 'handjob']
+from settings import IMG_PATH, TAGS, CLEAR
 
 INFO_KEY = ['id', 'file_ext', 'tags', 'file_url', 'author', 'creator_id']
 
@@ -43,6 +39,11 @@ class BaseYandeSpider:
         self.img_result = []
         self.urls = None
 
+        self.total_count = 0
+        self.created_img_count = 0
+        self.updated_img_count = 0
+        self.renamed_img_count = 0
+
         if not tags:
             self.tags = TAGS
         else:
@@ -61,8 +62,9 @@ class BaseYandeSpider:
         if failed:
             print("waiting to download failed urls")
             for i in range(30):
-                print(f"\r{30 - i}", end='')
+                print(f"{CLEAR}{30 - i}", end='')
                 time.sleep(1)
+            print('')
         img_infos = []
         self.failed_urls = []
         rs = (grequests.get(url, headers=self.headers, proxies=self.proxies,
@@ -88,32 +90,20 @@ class BaseYandeSpider:
     def run(self):
         self.urls = set(self.refresh())
         self.img_result = self.request_infos(self.urls)
+        if not self.img_result:
+            return False  # no image to download
         self.img_result = {item['id']: item for item in self.img_result}.values()
         # TODO 觉得没啥问题就把下面这个这删了吧
         assert len(set(info['file_url'] for info in self.img_result)) == len(self.img_result)
-
-        img_exists_list = []
-        img_not_exists_list = []
+        self.total_count += len(self.img_result)
 
         for info in self.img_result:
-            exists = self.process_update_info(info)
-            if not exists:
-                img_not_exists_list.append(info['file_url'])
-            else:
-                img_exists_list.append(info['file_url'])
+            self.process_update_info(info)
 
-        img_not_exists_str = "\n".join(img_not_exists_list)
-        self.log_fp.write(img_not_exists_str)
-        self.log_fp.write('\n--- images below already exists ---\n')
-        self.log_fp.write('\n'.join(img_exists_list))
-
-        print("total: %d, exists: %d, not exists: %d" %
-              (len(self.img_result), len(img_exists_list), len(img_not_exists_list)))
-
-        self.log_fp.close()
-
-    @staticmethod
-    def process_update_info(info):
+        print("total: %d, create: %d, update: %d include %d renamed images" %
+              (self.total_count, self.created_img_count, self.updated_img_count, self.renamed_img_count))
+        return True
+    def process_update_info(self, info):
         assert info['id']
         check_res = check_exists(Image, id=info['id'])
         rt = bool(check_res)
@@ -131,17 +121,44 @@ class BaseYandeSpider:
                 setattr(img, key, info[key])
                 diff = True
 
+        new_name = f'{img.id} {img.author}.{info["file_ext"]}'
+        new_name = new_name.translate(str.maketrans(r'/\:*?"<>|', "_________"))
+        if img.name != new_name:
+            diff = True
+
+        if not rt:
+            img.name = new_name
+            self.log_fp.write(f'create "{new_name}"\n')
+            self.created_img_count += 1
+            ss.commit()
+            return
+
         if diff:
-            new_name = f'yande.re {img.id} {img.tags}.{info["file_ext"]}'
-            old_path = os.path.join(IMG_PATH, img.name)
-            new_path = os.path.join(IMG_PATH, new_name)
-            assert os.path.exists(old_path)
-            if old_path != new_path:
-                os.rename(old_path, new_path)
+            assert img.name
+            self.log_fp.write(f'update :{img.name}"\n')
+            self.updated_img_count += 1
+
+            if img.name != new_name:
+                old_path = os.path.join(IMG_PATH, img.name)
+                new_path = os.path.join(IMG_PATH, new_name)
                 img.name = new_name
 
+                if not os.path.exists(old_path):
+                    if img.star >= 0:
+                        msg = f'file not found "{img.name}"\n'
+                        print(msg, end='')
+                        self.log_fp.write(msg)
+                        img.star = -2
+                        ss.commit()
+                    return
+
+                if old_path != new_path:
+                    self.log_fp.write(f'rename to "{new_name}"\n')
+                    self.renamed_img_count += 1
+                    os.rename(old_path, new_path)
+                    img.name = new_name
+
         ss.commit()
-        return rt
 
     def handle(self, req, info):
         print(f"{req.url} 请求失败")
@@ -150,6 +167,9 @@ class BaseYandeSpider:
     @staticmethod
     def request_callback(req, *args, **kwargs):
         print(f"{req.url}请求成功")
+
+    def __del__(self):
+        self.log_fp.close()
 
 
 class YandeDaily(BaseYandeSpider):
@@ -188,16 +208,23 @@ class YandeDaily(BaseYandeSpider):
 class YandeAll(BaseYandeSpider):
     METHOD = 'ALL'
 
+    def __init__(self, *args, **kwargs):
+        super(YandeAll, self).__init__(*args, **kwargs)
+        self.page = 1
+        self.batch_size = 3
+
     def refresh(self):
         tag2dl = self.tags[0]
-        print("%s?limit=1000&tags=%s&page=%s" % (self.post_url, tag2dl, 1))
-        pages = int(input('with a limit of 1000, how many pages?: '))
+        # print("%s?limit=1000&tags=%s&page=%s" % (self.post_url, tag2dl, 1))
+        # pages = int(input('with a limit of 1000, how many pages?: '))
 
-        for i in range(1, pages + 1):
+        for i in range(self.page, self.page + self.batch_size):
+            self.page += 1
             yield "%s?limit=1000&tags=%s&page=%s" % (self.post_url, tag2dl, str(i))
 
     def run(self):
-        super(YandeAll, self).run()
+        while super(YandeAll, self).run():
+            print('-'*7)
 
 
 # TODO: 如果请求太多的话还是分批比较好
@@ -214,15 +241,4 @@ class YandeId(BaseYandeSpider):
 
 
 if __name__ == '__main__':
-    typ = input("what kind of method you want: ")
-    if typ == 'daily':
-        yande = YandeDaily()
-    elif typ == 'all':
-        tag = input('the tag you want: ')
-        yande = YandeAll([tag])
-    elif typ == 'id':
-        yande = YandeId()
-    else:
-        print('unknown command.')
-        sys.exit(-1)
-    yande.run()
+    pass
